get location
