+++ app/src/main/res/layout/fragment_settings.xml
@@ -40,6 +40,11 @@
         android:id="@+id/buttonToGatLocation"
         />
 
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/tvLocation" />
+
     <ImageView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
